# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM Pipeline - Locação de Contas de Anúncio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Clients
...
  /integrations/whatsapp/leads:
    post:
      operationId: whatsappCreateLead
      tags: [integrations]
      summary: Create lead from WhatsApp
      security:
        - integrationKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappLeadInput"
      responses:
        "200":
          description: Client and deal created/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappLeadResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /integrations/whatsapp/deals/update-stage:
    post:
      operationId: whatsappUpdateDealStage
      tags: [integrations]
      summary: Update deal stage from WhatsApp
      security:
        - integrationKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappStageUpdate"
      responses:
        "200":
          description: Deal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /integrations/whatsapp/clients/update:
    post:
      operationId: whatsappUpdateClient
      tags: [integrations]
      summary: Update client data from WhatsApp
      security:
        - integrationKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappClientUpdate"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /integrations/whatsapp/client-by-phone:
    get:
      operationId: whatsappGetClientByPhone
      tags: [integrations]
      summary: Get client and deals by phone number
      security:
        - integrationKey: []
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client with deals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappClientLookup"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    integrationKey:
      type: apiKey
      in: header
      name: X-Integration-Key
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, sales]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string

        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, sales]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string

        role:
          type: string
          enum: [admin, sales]
        active:
          type: boolean
        password:
          type: string
          minLength: 6

    # Clients
    Client:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        document:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Deal:
      type: object
      required: [id, title, stage, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        ownerId:
          type: ["integer", "null"]
        stage:
          type: string
          enum: [lead_captado, qualificacao, proposta, negociacao, fechamento, onboarding, ativo, renovacao, encerrado]
        platform:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        niche:
          type: ["string", "null"]
        riskLevel:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        hasBlockHistory:
          type: ["boolean", "null"]
        blockDescription:
          type: ["string", "null"]
        proposalValue:
          type: ["number", "null"]
        proposalConditions:
          type: ["string", "null"]
        proposalTerm:
          type: ["string", "null"]
        acceptanceDate:
          type: ["string", "null"]
          format: date
        paymentMethod:
          type: ["string", "null"]
        termsAccepted:
          type: ["boolean", "null"]
        adAccountId:
          type: ["integer", "null"]
        adAccountCreated:
          type: ["boolean", "null"]
        adAccountLimit:
          type: ["number", "null"]
        firstCampaignActive:
          type: ["boolean", "null"]
        rentalStatus:
          type: ["string", "null"]
        contractEndDate:
          type: ["string", "null"]
          format: date
        clientName:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    WhatsappLeadInput:
      type: object
      required: [name]
      properties:
        externalId:
          type: string
        name:
          type: string
          minLength: 1
        phone:
          type: string
        niche:
          type: string
        platforms:
          type: array
          items:
            type: string
        estimatedInvestment:
          type: number
        hasBlockHistory:
          type: boolean
        blockDescription:
          type: string
        intent:
          type: string
    WhatsappLeadResponse:
      type: object
      required: [client, deal]
      properties:
        client:
          $ref: "#/components/schemas/Client"
        deal:
          $ref: "#/components/schemas/Deal"
    WhatsappStageUpdate:
      type: object
      required: [dealId, newStage]
      properties:
        dealId:
          type: integer
        newStage:
          type: string
          enum: [lead_captado, qualificacao, proposta, negociacao, fechamento, onboarding, ativo, renovacao, encerrado]
        reason:
          type: string
        trigger:
          type: string
    WhatsappClientUpdate:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          minLength: 1
        name:
          type: string
        niche:
          type: string
        hasBlockHistory:
          type: boolean
        blockDescription:
          type: string
        status:
          type: string
    WhatsappClientLookup:
      type: object
      required: [client, deals]
      properties:
        client:
          $ref: "#/components/schemas/Client"
        deals:
          type: array
          items:
            $ref: "#/components/schemas/Deal"