# lib/api-zod/src/generated/types/whatsappLeadInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM Pipeline - Locação de Contas de Anúncio
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsappLeadInput {
  externalId?: string;
  /** @minLength 1 */
  name: string;
  phone?: string;
  niche?: string;
  platforms?: string[];
  estimatedInvestment?: number;
  hasBlockHistory?: boolean;
  blockDescription?: string;
  intent?: string;
}