# lib/api-zod/src/generated/types/whatsappLeadResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * CRM Pipeline - Locação de Contas de Anúncio
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from './client';
import type { Deal } from './deal';

export interface WhatsappLeadResponse {
  client: Client;
  deal: Deal;
}